requirements.txt
python-dotenv
pypdfium2
langchain-text-splitters
chromadb
openai
colorama
pytest